% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhehaoColor.R
\name{zhehaoColor}
\alias{zhehaoColor}
\title{Generate Palette}
\usage{
zhehaoColor(palette_id, n)
}
\arguments{
\item{palette_id}{Name of Palette. Complete list of acceptable IDs can be found on Github.}

\item{n}{Number of Colors. If not specified, original palette will return, if n exceeds original palette, colors are automatically interpolated.}
}
\description{
Color palettes extracted from my \href{https://zzzhehao.github.io/gallery/gallery.html}{photographs}.
}
