% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{zhehaoplot_usePalette}
\alias{zhehaoplot_usePalette}
\title{Use ZhehaoColor Palette in ggplor Object.}
\usage{
zhehaoplot_usePalette(clp = "copenhagen1", n)
}
\description{
Use ZhehaoColor Palette in ggplor Object.
}
